Move OnMembersChanged to subtensor common
