+++ pallets/multi-collective/Cargo.toml
@@ -21,6 +21,7 @@ frame-system = { workspace = true }
 frame-support = { workspace = true }
 impl-trait-for-tuples = { workspace = true }
 num-traits = { workspace = true }
+subtensor-runtime-common = { workspace = true }
 
 [dev-dependencies]
 sp-io = { workspace = true, default-features = true }
@@ -35,11 +36,13 @@ std = [
     "frame-system/std",
     "frame-support/std",
     "num-traits/std",
+    "subtensor-runtime-common/std",
 ]
 runtime-benchmarks = [
 	"frame-support/runtime-benchmarks",
 	"frame-system/runtime-benchmarks",
-	"sp-runtime/runtime-benchmarks"
+	"sp-runtime/runtime-benchmarks",
+	"subtensor-runtime-common/runtime-benchmarks",
 ]
 try-runtime = [
 	"frame-support/try-runtime",
